Move quantifier methods to Quantifiers.cs
Refactored quantifier-related methods from Rejigs.cs into a new Quantifiers.cs file for better separation of concerns. Renamed the test class to QuantifiersTests to match conventions.
